Re-adding deleted classes to avoid breaking changes
Since we have solutions and packages using a mix of different versions of this package, it's not possible to upgrade to alpha007 without some other part breaking. Therefore the deleted classes have been re-added to the package, but marked as obsolete. The classes ought still to be removed before pushing a stable release.